+++ tag.http
@@ -1,7 +1,6 @@
-POST http://localhost:8080/admin/1/tags HTTP/1.1
+POST http://localhost:8080/admins/1/tags HTTP/1.1
 Content-Type: application/json
 
 {
-  "adminId": "1",
   "title": "고풍스러운"
 }
